azure_api/requirements.txt
-r ../app/requirements.txt
azure-functions
